requirements.txt
zeroc-ice==3.6.5
omero-py==5.8.3
pandas==1.1.4
numpy==1.19.4
xlrd==1.2.0